Unify implementation of read_size
Modernisations applied in headernt.c but not applied to header.c - use
uint32_t rather than unsigned long

+++ stdlib/headernt.c
@@ -53,11 +53,11 @@ static void write_console(HANDLE hOut, WCHAR *wstr)
   }
 }
 
-static unsigned long read_size(const char * const ptr)
+static uint32_t read_size(const char * const ptr)
 {
   const unsigned char * const p = (const unsigned char * const) ptr;
-  return ((unsigned long) p[0] << 24) | ((unsigned long) p[1] << 16) |
-         ((unsigned long) p[2] << 8) | p[3];
+  return ((uint32_t) p[0] << 24) | ((uint32_t) p[1] << 16) |
+         ((uint32_t) p[2] << 8) | p[3];
 }
 
 static char * read_runtime_path(HANDLE h)
